+++ packages/core/CHANGELOG.md
@@ -162,7 +162,7 @@
 
   ```ts
   export default defineConfig({
-    reporters: ["file"],
+    reporters: ['file'],
   });
   ```
 
@@ -178,7 +178,7 @@
 
   ```ts
   export default defineConfig({
-    reporters: ["default", "file"],
+    reporters: ['default', 'file'],
   });
   ```